refactor:  muda arquivo page da página inicial para arquivo de grupo de pedidos.
